+++ XMonad/Layout/WorkspaceDir.hs
@@ -58,6 +58,11 @@ import XMonad.StackSet ( tag, currentTag )
 --
 -- >  , ((modm .|. shiftMask, xK_x     ), changeDir def)
 --
+-- If you prefer a prompt with case-insensitive completion:
+--
+-- >  , ((modm .|. shiftMask, xK_x     ),
+--       changeDir def {complCaseSensitivity = ComplCaseSensitive False})
+--
 -- For detailed instruction on editing the key binding see:
 --
 -- "XMonad.Doc.Extending#Editing_key_bindings".